Assign number of points per question.
Teachers can assign different number of points per question.

+++ src/quiz/entities/quiz-result.entity.ts
@@ -26,11 +26,11 @@ export class QuizResult {
   @Field(() => User)
   author: User;
 
-  @Column('int')
+  @Column('float')
   @Field(() => Number)
   score: number;
 
-  @Column('int')
+  @Column('float')
   @Field(() => Number)
   total: number;
 }
